refactored generic examples
