+++ _content/tour/generics-basics.article
@@ -1,15 +1,21 @@
-Functions
-Learn how to use generics in Go.
+Basics
+Learn how to write a basic generic print function.
 
-* Generics - Functions
+* Generics - Basics
 
-Learn how to use generics in Go.
+Learn how to write a basic generic print function.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Concrete implementation of print
+- *Example* *2*: Type Assertion implementation of print
+- *Example* *3*: Reflection implementation of print
+- *Example* *4*: Generic implementation of print
 
-.play generics/functions/example1.go
+.play generics/basics/example1.go
+.play generics/basics/example2.go
+.play generics/basics/example3.go
+.play generics/basics/example4.go
 
 ** Explained
