+++ _content/tour/generics-behavior-constraints.article
@@ -10,9 +10,15 @@ predeclared constraint identifier any.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Concrete stringify function
+- *Example* *2*: Type assertion stringify function
+- *Example* *3*: Interface stringify function
+- *Example* *4*: Generic stringify function
 
 .play generics/behavior-constraints/example1.go
+.play generics/behavior-constraints/example2.go
+.play generics/behavior-constraints/example3.go
+.play generics/behavior-constraints/example4.go
 
 ** Explained
