+++ _content/tour/generics-channels.article
@@ -7,9 +7,11 @@ Explore how the Go team could add a package of concurrency patterns into the sta
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Work Function
+- *Example* *2*: Pooling
 
 .play generics/channels/example1.go
+.play generics/channels/example2.go
 
 ** Explained
