+++ _content/tour/generics-hash-table.article
@@ -9,7 +9,7 @@ he wrote. It's a great example of what is possible with generics.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Hash Table
 
 .play generics/hash-table/example1.go
