+++ _content/tour/generics-multi-type-params.article
@@ -7,7 +7,7 @@ You're not restricted to using just one generic type at a time.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Print function
 
 .play generics/multi-type-params/example1.go
