+++ _content/tour/generics-slice-constraints.article
@@ -1,13 +1,13 @@
 Slice Constraints
-There may be times where you need to define a user-defined slice based on an underlying type.
+There may be times where you need to constrain the generic type to be only a slice.
 
 * Generics - Slice Constraints
 
-There may be times where you need to define a user-defined slice based on an underlying type.
+There may be times where you need to constrain the generic type to be only a slice.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Operate function
 
 .play generics/slice-constraints/example1.go
