+++ _content/tour/generics-struct-types.article
@@ -1,4 +1,4 @@
-Declare Generic Type Using Struct Types
+Struct Types
 You can declare a generic type using a struct type.
 
 * Generics - Struct Types
@@ -7,7 +7,7 @@ You can declare a generic type using a struct type.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Linked List
 
 .play generics/struct-types/example1.go
