+++ _content/tour/generics-type-constraints.article
@@ -8,9 +8,13 @@ This only works for generics.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Type based constraint
+- *Example* *2*: Predeclared type constraint "comparable"
+- *Example* *3*: Mix type and behavior constraints
 
 .play generics/type-constraints/example1.go
+.play generics/type-constraints/example2.go
+.play generics/type-constraints/example3.go
 
 ** Explained
