+++ _content/tour/generics-underlying-types.article
@@ -1,4 +1,4 @@
-Declare Generic Type Using Underlying Types
+Underlying Types
 You can declare a generic type using an underlying type.
 
 * Generics - Underlying Types
@@ -7,9 +7,13 @@ You can declare a generic type using an underlying type.
 
 ** Code Review
 
-- *Example* *1*: Basic Example
+- *Example* *1*: Concrete vector type
+- *Example* *2*: Interface vector type
+- *Example* *3*: Generic vector type
 
 .play generics/underlying-types/example1.go
+.play generics/underlying-types/example2.go
+.play generics/underlying-types/example3.go
 
 ** Explained
