+++ _content/tour/generics/behavior-constraints/example2.go
@@ -4,30 +4,34 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// Sample program to show how to implement a stringify function that is
-// specific to each of the concrete types implemented above. In each case,
-// the stringify function returns a slice of strings. These function use
-// the String method against the individual user or customer value.
+// Sample program to show how to implement an empty interface solution which
+// uses type assertions for the different concrete slices to be supported.
+// We've basically moved the functions from above into case statements.
+// This function uses the String method against the value.
 package main
 
 import (
 	"fmt"
 )
 
-func stringifyUsers(users []user) []string {
-	ret := make([]string, 0, len(users))
-	for _, user := range users {
-		ret = append(ret, user.String())
-	}
-	return ret
-}
+func stringifyAssert(v interface{}) []string {
+	switch list := v.(type) {
+	case []user:
+		ret := make([]string, 0, len(list))
+		for _, value := range list {
+			ret = append(ret, value.String())
+		}
+		return ret
 
-func stringifyCustomers(customers []customer) []string {
-	ret := make([]string, 0, len(customers))
-	for _, customer := range customers {
-		ret = append(ret, customer.String())
+	case []customer:
+		ret := make([]string, 0, len(list))
+		for _, value := range list {
+			ret = append(ret, value.String())
+		}
+		return ret
 	}
-	return ret
+
+	return nil
 }
 
 // Defining two types that implement the fmt.Stringer interface. Each
@@ -58,14 +62,14 @@ func main() {
 		{name: "Bill", email: "bill@ardanlabs.com"},
 		{name: "Ale", email: "ale@whatever.com"},
 	}
-	s1 := stringifyUsers(users)
+	s1 := stringifyAssert(users)
 
 	customers := []customer{
 		{name: "Google", email: "you@google.com"},
 		{name: "MSFT", email: "you@msft.com"},
 	}
-	s2 := stringifyCustomers(customers)
+	s2 := stringifyAssert(customers)
 
-	fmt.Println("users Con:", s1)
-	fmt.Println("cust Con:", s2)
+	fmt.Println("users Int:", s1)
+	fmt.Println("cust Int:", s2)
 }
