+++ _content/tour/generics/behavior-constraints/example4.go
@@ -4,29 +4,25 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// Sample program to show how to implement a stringify function that is
-// specific to each of the concrete types implemented above. In each case,
-// the stringify function returns a slice of strings. These function use
-// the String method against the individual user or customer value.
+// Sample program to show how to implement a generics solution which allows
+// a slice of some type T (to be determined later) to be passed and stringified.
+// This code more closely resembles the concrete implementations that we started
+// with and is easier to read than the reflect implementation. However, an
+// interface constraint of type fmt.Stringer is applied to allow the compiler
+// to know the value of type T passed requires a String method.
 package main
 
 import (
 	"fmt"
 )
 
-func stringifyUsers(users []user) []string {
-	ret := make([]string, 0, len(users))
-	for _, user := range users {
-		ret = append(ret, user.String())
-	}
-	return ret
-}
+func stringify[T fmt.Stringer](slice []T) []string {
+	ret := make([]string, 0, len(slice))
 
-func stringifyCustomers(customers []customer) []string {
-	ret := make([]string, 0, len(customers))
-	for _, customer := range customers {
-		ret = append(ret, customer.String())
+	for _, value := range slice {
+		ret = append(ret, value.String())
 	}
+
 	return ret
 }
 
@@ -58,14 +54,14 @@ func main() {
 		{name: "Bill", email: "bill@ardanlabs.com"},
 		{name: "Ale", email: "ale@whatever.com"},
 	}
-	s1 := stringifyUsers(users)
+	s1 := stringify(users)
 
 	customers := []customer{
 		{name: "Google", email: "you@google.com"},
 		{name: "MSFT", email: "you@msft.com"},
 	}
-	s2 := stringifyCustomers(customers)
+	s2 := stringify(customers)
 
-	fmt.Println("users Con:", s1)
-	fmt.Println("cust Con:", s2)
+	fmt.Println("users Gen:", s1)
+	fmt.Println("cust Gen:", s2)
 }
