+++ _content/tour/generics/channels/example1.go
@@ -1,18 +1,20 @@
+//go:build OMIT
+// +build OMIT
+
+// All material is licensed under the Apache License Version 2.0, January 2004
+// http://www.apache.org/licenses/LICENSE-2.0
+
+// Sample program to show how to execute a work function in a goroutine and
+// return a channel of type Result (to be determined later) back to the caller.
 package main
 
 import (
 	"context"
 	"fmt"
 	"math/rand"
-	"runtime"
-	"sync"
 	"time"
 )
 
-// =============================================================================
-
-// doWork will execute a work function in a goroutine and return a channel of
-// type Result (to be determined later) back to the caller.
 type doworkFn[Result any] func(context.Context) Result
 
 func doWork[Result any](ctx context.Context, work doworkFn[Result]) chan Result {
@@ -26,42 +28,8 @@ func doWork[Result any](ctx context.Context, work doworkFn[Result]) chan Result 
 	return ch
 }
 
-// =============================================================================
-
-// poolWork will execute a work function via a pool of goroutines and return a
-// channel of type Input (to be determined later) back to the caller. Once input
-// is received by any given goroutine, the work function is executed and the
-// Result value is displayed.
-type poolWorkFn[Input any, Result any] func(input Input) Result
-
-func poolWork[Input any, Result any](size int, work poolWorkFn[Input, Result]) (chan Input, func()) {
-	var wg sync.WaitGroup
-	wg.Add(size)
-
-	ch := make(chan Input)
-
-	for i := 0; i < size; i++ {
-		go func() {
-			defer wg.Done()
-			for input := range ch {
-				result := work(input)
-				fmt.Println("pollWork :", result)
-			}
-		}()
-	}
-
-	cancel := func() {
-		close(ch)
-		wg.Wait()
-	}
-
-	return ch, cancel
-}
-
-// =============================================================================
-
 func main() {
-	ctx, cancel := context.WithTimeout(context.Background(), 100 * time.Millisecond)
+	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
 	defer cancel()
 
 	dwf := func(ctx context.Context) string {
@@ -74,20 +42,5 @@ func main() {
 		fmt.Println("main:", v)
 	case <-ctx.Done():
 		fmt.Println("main: timeout")
-	}
-
-	fmt.Println("-------------------------------------------------------------")
-
-	size := runtime.GOMAXPROCS(0)
-	pwf := func(input int) string {
-		time.Sleep(time.Duration(rand.Intn(200)) * time.Millisecond)
-		return fmt.Sprintf("%d : received", input)
-	}
-	
-	ch, cancel := poolWork(size, pwf)
-	defer cancel()
-	
-	for i := 0; i < 5; i++ {
-		ch <- i
 	}
 }
