--- _content/tour/generics/io/example1.go
@@ -1,799 +1,0 @@
-// This code is provided by Rodger Peppe (@rogpeppe)
-// https://twitter.com/rogpeppe/status/1301189885836001281?s=20
-
-package main
-
-import (
-	"errors"
-	"fmt"
-	"io"
-	"sync"
-)
-
-func main() {
-	r, w := Pipe[int]()
-	go func() {
-		w.Write([]int{1, 2, 3, 4})
-		w.Write([]int{10, 11})
-		w.Close()
-	}()
-	buf := make([]int, 20)
-	n, _ := ReadFull[int](r, buf)
-	fmt.Println(buf[0:n])
-}
-
-// ==============================================================================
-
-// Seek whence values.
-const (
-	SeekStart   = 0 // seek relative to the origin of the file
-	SeekCurrent = 1 // seek relative to the current offset
-	SeekEnd     = 2 // seek relative to the end
-)
-
-// ErrShortWrite means that a write accepted fewer bytes than requested
-// but failed to return an explicit error.
-var ErrShortWrite = errors.New("short write")
-
-// ErrShortBuffer means that a read required a longer buffer than was provided.
-var ErrShortBuffer = errors.New("short buffer")
-
-// EOF is the error returned by Read when no more input is available.
-// Functions should return EOF only to signal a graceful end of input.
-// If the EOF occurs unexpectedly in a structured data stream,
-// the appropriate error is either ErrUnexpectedEOF or some other error
-// giving more detail.
-var EOF = io.EOF
-
-// ErrUnexpectedEOF means that EOF was encountered in the
-// middle of reading a fixed-size block or data structure.
-var ErrUnexpectedEOF = io.ErrUnexpectedEOF
-
-// ErrNoProgress is returned by some clients of an io.Reader when
-// many calls to Read have failed to return any data or error,
-// usually the sign of a broken io.Reader implementation.
-var ErrNoProgress = io.ErrNoProgress
-
-// Reader is the interface that wraps the basic Read method.
-//
-// Read reads up to len(p) bytes into p. It returns the number of bytes
-// read (0 <= n <= len(p)) and any error encountered. Even if Read
-// returns n < len(p), it may use all of p as scratch space during the call.
-// If some data is available but not len(p) bytes, Read conventionally
-// returns what is available instead of waiting for more.
-//
-// When Read encounters an error or end-of-file condition after
-// successfully reading n > 0 bytes, it returns the number of
-// bytes read. It may return the (non-nil) error from the same call
-// or return the error (and n == 0) from a subsequent call.
-// An instance of this general case is that a Reader returning
-// a non-zero number of bytes at the end of the input stream may
-// return either err == EOF or err == nil. The next Read should
-// return 0, EOF.
-//
-// Callers should always process the n > 0 bytes returned before
-// considering the error err. Doing so correctly handles I/O errors
-// that happen after reading some bytes and also both of the
-// allowed EOF behaviors.
-//
-// Implementations of Read are discouraged from returning a
-// zero byte count with a nil error, except when len(p) == 0.
-// Callers should treat a return of 0 and nil as indicating that
-// nothing happened; in particular it does not indicate EOF.
-//
-// Implementations must not retain p.
-type Reader[T any] interface {
-	Read(p []T) (n int, err error)
-}
-
-// Writer is the interface that wraps the basic Write method.
-//
-// Write writes len(p) bytes from p to the underlying data stream.
-// It returns the number of bytes written from p (0 <= n <= len(p))
-// and any error encountered that caused the write to stop early.
-// Write must return a non-nil error if it returns n < len(p).
-// Write must not modify the slice data, even temporarily.
-//
-// Implementations must not retain p.
-type Writer[T any] interface {
-	Write(p []T) (n int, err error)
-}
-
-// Closer is the interface that wraps the basic Close method.
-//
-// The behavior of Close after the first call is undefined.
-// Specific implementations may document their own behavior.
-type Closer interface {
-	Close() error
-}
-
-// Seeker is the interface that wraps the basic Seek method.
-//
-// Seek sets the offset for the next Read or Write to offset,
-// interpreted according to whence:
-// SeekStart means relative to the start of the file,
-// SeekCurrent means relative to the current offset, and
-// SeekEnd means relative to the end.
-// Seek returns the new offset relative to the start of the
-// file and an error, if any.
-//
-// Seeking to an offset before the start of the file is an error.
-// Seeking to any positive offset is legal, but the behavior of subsequent
-// I/O operations on the underlying object is implementation-dependent.
-type Seeker interface {
-	Seek(offset int64, whence int) (int64, error)
-}
-
-// ReadWriter is the interface that groups the basic Read and Write methods.
-type ReadWriter[T any] interface {
-	Reader[T]
-	Writer[T]
-}
-
-// ReadCloser is the interface that groups the basic Read and Close methods.
-type ReadCloser[T any] interface {
-	Reader[T]
-	Closer
-}
-
-// WriteCloser is the interface that groups the basic Write and Close methods.
-type WriteCloser[T any] interface {
-	Writer[T]
-	Closer
-}
-
-// ReadWriteCloser is the interface that groups the basic Read, Write and Close methods.
-type ReadWriteCloser[T any] interface {
-	Reader[T]
-	Writer[T]
-	Closer
-}
-
-// ReadSeeker is the interface that groups the basic Read and Seek methods.
-type ReadSeeker[T any] interface {
-	Reader[T]
-	Seeker
-}
-
-// WriteSeeker is the interface that groups the basic Write and Seek methods.
-type WriteSeeker[T any] interface {
-	Writer[T]
-	Seeker
-}
-
-// ReadWriteSeeker is the interface that groups the basic Read, Write and Seek methods.
-type ReadWriteSeeker[T any] interface {
-	Reader[T]
-	Writer[T]
-	Seeker
-}
-
-// ReaderFrom is the interface that wraps the ReadFrom method.
-//
-// ReadFrom reads data from r until EOF or error.
-// The return value n is the number of bytes read.
-// Any error except io.EOF encountered during the read is also returned.
-//
-// The Copy function uses ReaderFrom if available.
-type ReaderFrom[T any] interface {
-	ReadFrom(r Reader[T]) (n int64, err error)
-}
-
-// WriterTo is the interface that wraps the WriteTo method.
-//
-// WriteTo writes data to w until there's no more data to write or
-// when an error occurs. The return value n is the number of bytes
-// written. Any error encountered during the write is also returned.
-//
-// The Copy function uses WriterTo if available.
-type WriterTo[T any] interface {
-	WriteTo(w Writer[T]) (n int64, err error)
-}
-
-// ReaderAt is the interface that wraps the basic ReadAt method.
-//
-// ReadAt reads len(p) bytes into p starting at offset off in the
-// underlying input source. It returns the number of bytes
-// read (0 <= n <= len(p)) and any error encountered.
-//
-// When ReadAt returns n < len(p), it returns a non-nil error
-// explaining why more bytes were not returned. In this respect,
-// ReadAt is stricter than Read.
-//
-// Even if ReadAt returns n < len(p), it may use all of p as scratch
-// space during the call. If some data is available but not len(p) bytes,
-// ReadAt blocks until either all the data is available or an error occurs.
-// In this respect ReadAt is different from Read.
-//
-// If the n = len(p) bytes returned by ReadAt are at the end of the
-// input source, ReadAt may return either err == EOF or err == nil.
-//
-// If ReadAt is reading from an input source with a seek offset,
-// ReadAt should not affect nor be affected by the underlying
-// seek offset.
-//
-// Clients of ReadAt can execute parallel ReadAt calls on the
-// same input source.
-//
-// Implementations must not retain p.
-type ReaderAt[T any] interface {
-	ReadAt(p []T, off int64) (n int, err error)
-}
-
-// WriterAt is the interface that wraps the basic WriteAt method.
-//
-// WriteAt writes len(p) bytes from p to the underlying data stream
-// at offset off. It returns the number of bytes written from p (0 <= n <= len(p))
-// and any error encountered that caused the write to stop early.
-// WriteAt must return a non-nil error if it returns n < len(p).
-//
-// If WriteAt is writing to a destination with a seek offset,
-// WriteAt should not affect nor be affected by the underlying
-// seek offset.
-//
-// Clients of WriteAt can execute parallel WriteAt calls on the same
-// destination if the ranges do not overlap.
-//
-// Implementations must not retain p.
-type WriterAt[T any] interface {
-	WriteAt(p []T, off int64) (n int, err error)
-}
-
-// ItemReader is the interface that wraps the ReadItem method.
-//
-// ReadItem reads and returns the next byte from the input or
-// any error encountered. If ReadItem returns an error, no input
-// was consumed, and the returned value is undefined.
-//
-// ReadItem provides an efficient interface for item-at-time
-// processing. A Reader that does not implement ItemReader
-// can be wrapped using genericbufio.NewReader to add this method.
-type ItemReader[T any] interface {
-	ReadItem() (T, error)
-}
-
-// ItemScanner is the interface that adds the UnreadItem method to the
-// basic ReadItem method.
-//
-// UnreadItem causes the next call to ReadItem to return the same byte
-// as the previous call to ReadItem.
-// It may be an error to call UnreadItem twice without an intervening
-// call to ReadItem.
-type ItemScanner[T any] interface {
-	ItemReader[T]
-	UnreadItem() error
-}
-
-// ItemWriter is the interface that wraps the WriteItem method.
-type ItemWriter[T any] interface {
-	WriteItem(c T) error
-}
-
-// ReadAtLeast reads from r into buf until it has read at least min bytes.
-// It returns the number of bytes copied and an error if fewer bytes were read.
-// The error is EOF only if no bytes were read.
-// If an EOF happens after reading fewer than min bytes,
-// ReadAtLeast returns ErrUnexpectedEOF.
-// If min is greater than the length of buf, ReadAtLeast returns ErrShortBuffer.
-// On return, n >= min if and only if err == nil.
-// If r returns an error having read at least min bytes, the error is dropped.
-func ReadAtLeast[T any](r Reader[T], buf []T, min int) (n int, err error) {
-	if len(buf) < min {
-		return 0, ErrShortBuffer
-	}
-	for n < min && err == nil {
-		var nn int
-		nn, err = r.Read(buf[n:])
-		n += nn
-	}
-	if n >= min {
-		err = nil
-	} else if n > 0 && err == EOF {
-		err = ErrUnexpectedEOF
-	}
-	return
-}
-
-// ReadFull reads exactly len(buf) bytes from r into buf.
-// It returns the number of bytes copied and an error if fewer bytes were read.
-// The error is EOF only if no bytes were read.
-// If an EOF happens after reading some but not all the bytes,
-// ReadFull returns ErrUnexpectedEOF.
-// On return, n == len(buf) if and only if err == nil.
-// If r returns an error having read at least len(buf) bytes, the error is dropped.
-func ReadFull[T any](r Reader[T], buf []T) (n int, err error) {
-	return ReadAtLeast(r, buf, len(buf))
-}
-
-// CopyN copies n bytes (or until an error) from src to dst.
-// It returns the number of bytes copied and the earliest
-// error encountered while copying.
-// On return, written == n if and only if err == nil.
-//
-// If dst implements the ReaderFrom interface,
-// the copy is implemented using it.
-func CopyN[T any](dst Writer[T], src Reader[T], n int64) (written int64, err error) {
-	written, err = Copy(dst, LimitReader(src, n))
-	if written == n {
-		return n, nil
-	}
-	if written < n && err == nil {
-		// src stopped early; must have been EOF.
-		err = EOF
-	}
-	return
-}
-
-// Copy copies from src to dst until either EOF is reached
-// on src or an error occurs. It returns the number of bytes
-// copied and the first error encountered while copying, if any.
-//
-// A successful Copy returns err == nil, not err == EOF.
-// Because Copy is defined to read from src until EOF, it does
-// not treat an EOF from Read as an error to be reported.
-//
-// If src implements the WriterTo interface,
-// the copy is implemented by calling src.WriteTo(dst).
-// Otherwise, if dst implements the ReaderFrom interface,
-// the copy is implemented by calling dst.ReadFrom(src).
-func Copy[T any](dst Writer[T], src Reader[T]) (written int64, err error) {
-	return copyBuffer(dst, src, nil)
-}
-
-// CopyBuffer is identical to Copy except that it stages through the
-// provided buffer (if one is required) rather than allocating a
-// temporary one. If buf is nil, one is allocated; otherwise if it has
-// zero length, CopyBuffer panics.
-//
-// If either src implements WriterTo or dst implements ReaderFrom,
-// buf will not be used to perform the copy.
-func CopyBuffer[T any](dst Writer[T], src Reader[T], buf []T) (written int64, err error) {
-	if buf != nil && len(buf) == 0 {
-		panic("empty buffer in io.CopyBuffer")
-	}
-	return copyBuffer(dst, src, buf)
-}
-
-// copyBuffer is the actual implementation of Copy and CopyBuffer.
-// if buf is nil, one is allocated.
-func copyBuffer[T any](dst Writer[T], src Reader[T], buf []T) (written int64, err error) {
-	// If the reader has a WriteTo method, use it to do the copy.
-	// Avoids an allocation and a copy.
-	if wt, ok := src.(WriterTo[T]); ok {
-		return wt.WriteTo(dst)
-	}
-	// Similarly, if the writer has a ReadFrom method, use it to do the copy.
-	if rt, ok := dst.(ReaderFrom[T]); ok {
-		return rt.ReadFrom(src)
-	}
-	if buf == nil {
-		size := 32 * 1024
-		if l, ok := src.(*LimitedReader[T]); ok && int64(size) > l.N {
-			if l.N < 1 {
-				size = 1
-			} else {
-				size = int(l.N)
-			}
-		}
-		buf = make([]T, size)
-	}
-	for {
-		nr, er := src.Read(buf)
-		if nr > 0 {
-			nw, ew := dst.Write(buf[0:nr])
-			if nw > 0 {
-				written += int64(nw)
-			}
-			if ew != nil {
-				err = ew
-				break
-			}
-			if nr != nw {
-				err = ErrShortWrite
-				break
-			}
-		}
-		if er != nil {
-			if er != EOF {
-				err = er
-			}
-			break
-		}
-	}
-	return written, err
-}
-
-// LimitReader returns a Reader that reads from r
-// but stops with EOF after n bytes.
-// The underlying implementation is a *LimitedReader.
-func LimitReader[T any](r Reader[T], n int64) Reader[T] { return &LimitedReader[T]{r, n} }
-
-// A LimitedReader reads from R but limits the amount of
-// data returned to just N bytes. Each call to Read
-// updates N to reflect the new amount remaining.
-// Read returns EOF when N <= 0 or when the underlying R returns EOF.
-type LimitedReader[T any] struct {
-	R Reader[T] // underlying reader
-	N int64     // max bytes remaining
-}
-
-func (l *LimitedReader[T]) Read(p []T) (n int, err error) {
-	if l.N <= 0 {
-		return 0, EOF
-	}
-	if int64(len(p)) > l.N {
-		p = p[0:l.N]
-	}
-	n, err = l.R.Read(p)
-	l.N -= int64(n)
-	return
-}
-
-// NewSectionReader returns a SectionReader that reads from r
-// starting at offset off and stops with EOF after n bytes.
-func NewSectionReader[T any](r ReaderAt[T], off int64, n int64) *SectionReader[T] {
-	return &SectionReader[T]{r, off, off, off + n}
-}
-
-// SectionReader implements Read, Seek, and ReadAt on a section
-// of an underlying ReaderAt.
-type SectionReader[T any] struct {
-	r     ReaderAt[T]
-	base  int64
-	off   int64
-	limit int64
-}
-
-func (s *SectionReader[T]) Read(p []T) (n int, err error) {
-	if s.off >= s.limit {
-		return 0, EOF
-	}
-	if max := s.limit - s.off; int64(len(p)) > max {
-		p = p[0:max]
-	}
-	n, err = s.r.ReadAt(p, s.off)
-	s.off += int64(n)
-	return
-}
-
-var errWhence = errors.New("Seek: invalid whence")
-var errOffset = errors.New("Seek: invalid offset")
-
-func (s *SectionReader[T]) Seek(offset int64, whence int) (int64, error) {
-	switch whence {
-	default:
-		return 0, errWhence
-	case SeekStart:
-		offset += s.base
-	case SeekCurrent:
-		offset += s.off
-	case SeekEnd:
-		offset += s.limit
-	}
-	if offset < s.base {
-		return 0, errOffset
-	}
-	s.off = offset
-	return offset - s.base, nil
-}
-
-func (s *SectionReader[T]) ReadAt(p []T, off int64) (n int, err error) {
-	if off < 0 || off >= s.limit-s.base {
-		return 0, EOF
-	}
-	off += s.base
-	if max := s.limit - off; int64(len(p)) > max {
-		p = p[0:max]
-		n, err = s.r.ReadAt(p, off)
-		if err == nil {
-			err = EOF
-		}
-		return n, err
-	}
-	return s.r.ReadAt(p, off)
-}
-
-// Size returns the size of the section in bytes.
-func (s *SectionReader[T]) Size() int64 { return s.limit - s.base }
-
-// TeeReader returns a Reader that writes to w what it reads from r.
-// All reads from r performed through it are matched with
-// corresponding writes to w. There is no internal buffering -
-// the write must complete before the read completes.
-// Any error encountered while writing is reported as a read error.
-func TeeReader[T any](r Reader[T], w Writer[T]) Reader[T] {
-	return &teeReader[T]{r, w}
-}
-
-type teeReader[T any] struct {
-	r Reader[T]
-	w Writer[T]
-}
-
-func (t *teeReader[T]) Read(p []T) (n int, err error) {
-	n, err = t.r.Read(p)
-	if n > 0 {
-		if n, err := t.w.Write(p[:n]); err != nil {
-			return n, err
-		}
-	}
-	return
-}
-
-// --------------------------- pipes
-
-// onceError is an object that will only store an error once.
-type onceError struct {
-	sync.Mutex // guards following
-	err        error
-}
-
-func (a *onceError) Store(err error) {
-	a.Lock()
-	defer a.Unlock()
-	if a.err != nil {
-		return
-	}
-	a.err = err
-}
-func (a *onceError) Load() error {
-	a.Lock()
-	defer a.Unlock()
-	return a.err
-}
-
-// ErrClosedPipe is the error used for read or write operations on a closed pipe.
-var ErrClosedPipe = errors.New("io: read/write on closed pipe")
-
-// A pipe is the shared pipe structure underlying PipeReader and PipeWriter.
-type pipe[T any] struct {
-	wrMu sync.Mutex // Serializes Write operations
-	wrCh chan []T
-	rdCh chan int
-
-	once sync.Once // Protects closing done
-	done chan struct{}
-	rerr onceError
-	werr onceError
-}
-
-func (p *pipe[T]) Read(b []T) (n int, err error) {
-	select {
-	case <-p.done:
-		return 0, p.readCloseError()
-	default:
-	}
-
-	select {
-	case bw := <-p.wrCh:
-		nr := copy(b, bw)
-		p.rdCh <- nr
-		return nr, nil
-	case <-p.done:
-		return 0, p.readCloseError()
-	}
-}
-
-func (p *pipe[T]) readCloseError() error {
-	rerr := p.rerr.Load()
-	if werr := p.werr.Load(); rerr == nil && werr != nil {
-		return werr
-	}
-	return ErrClosedPipe
-}
-
-func (p *pipe[T]) CloseRead(err error) error {
-	if err == nil {
-		err = ErrClosedPipe
-	}
-	p.rerr.Store(err)
-	p.once.Do(func() { close(p.done) })
-	return nil
-}
-
-func (p *pipe[T]) Write(b []T) (n int, err error) {
-	select {
-	case <-p.done:
-		return 0, p.writeCloseError()
-	default:
-		p.wrMu.Lock()
-		defer p.wrMu.Unlock()
-	}
-
-	for once := true; once || len(b) > 0; once = false {
-		select {
-		case p.wrCh <- b:
-			nw := <-p.rdCh
-			b = b[nw:]
-			n += nw
-		case <-p.done:
-			return n, p.writeCloseError()
-		}
-	}
-	return n, nil
-}
-
-func (p *pipe[T]) writeCloseError() error {
-	werr := p.werr.Load()
-	if rerr := p.rerr.Load(); werr == nil && rerr != nil {
-		return rerr
-	}
-	return ErrClosedPipe
-}
-
-func (p *pipe[T]) CloseWrite(err error) error {
-	if err == nil {
-		err = EOF
-	}
-	p.werr.Store(err)
-	p.once.Do(func() { close(p.done) })
-	return nil
-}
-
-// A PipeReader is the read half of a pipe.
-type PipeReader[T any] struct {
-	p *pipe[T]
-}
-
-// Read implements the standard Read interface:
-// it reads data from the pipe, blocking until a writer
-// arrives or the write end is closed.
-// If the write end is closed with an error, that error is
-// returned as err; otherwise err is EOF.
-func (r *PipeReader[T]) Read(data []T) (n int, err error) {
-	return r.p.Read(data)
-}
-
-// Close closes the reader; subsequent writes to the
-// write half of the pipe will return the error ErrClosedPipe.
-func (r *PipeReader[T]) Close() error {
-	return r.CloseWithError(nil)
-}
-
-// CloseWithError closes the reader; subsequent writes
-// to the write half of the pipe will return the error err.
-//
-// CloseWithError never overwrites the previous error if it exists
-// and always returns nil.
-func (r *PipeReader[T]) CloseWithError(err error) error {
-	return r.p.CloseRead(err)
-}
-
-// A PipeWriter is the write half of a pipe.
-type PipeWriter[T any] struct {
-	p *pipe[T]
-}
-
-// Write implements the standard Write interface:
-// it writes data to the pipe, blocking until one or more readers
-// have consumed all the data or the read end is closed.
-// If the read end is closed with an error, that err is
-// returned as err; otherwise err is ErrClosedPipe.
-func (w *PipeWriter[T]) Write(data []T) (n int, err error) {
-	return w.p.Write(data)
-}
-
-// Close closes the writer; subsequent reads from the
-// read half of the pipe will return no Ts and EOF.
-func (w *PipeWriter[T]) Close() error {
-	return w.CloseWithError(nil)
-}
-
-// CloseWithError closes the writer; subsequent reads from the
-// read half of the pipe will return no Ts and the error err,
-// or EOF if err is nil.
-//
-// CloseWithError never overwrites the previous error if it exists
-// and always returns nil.
-func (w *PipeWriter[T]) CloseWithError(err error) error {
-	return w.p.CloseWrite(err)
-}
-
-// Pipe creates a synchronous in-memory pipe.
-// It can be used to connect code expecting an io.Reader
-// with code expecting an io.Writer.
-//
-// Reads and Writes on the pipe are matched one to one
-// except when multiple Reads are needed to consume a single Write.
-// That is, each Write to the PipeWriter blocks until it has satisfied
-// one or more Reads from the PipeReader that fully consume
-// the written data.
-// The data is copied directly from the Write to the corresponding
-// Read (or Reads); there is no internal buffering.
-//
-// It is safe to call Read and Write in parallel with each other or with Close.
-// Parallel calls to Read and parallel calls to Write are also safe:
-// the individual calls will be gated sequentially.
-func Pipe[T any]() (*PipeReader[T], *PipeWriter[T]) {
-	p := &pipe[T]{
-		wrCh: make(chan []T),
-		rdCh: make(chan int),
-		done: make(chan struct{}),
-	}
-	return &PipeReader[T]{p}, &PipeWriter[T]{p}
-}
-
-// --------------------- multi
-
-type eofReader[T any] struct{}
-
-func (eofReader[T]) Read([]T) (int, error) {
-	return 0, EOF
-}
-
-type multiReader[T any] struct {
-	readers []Reader[T]
-}
-
-func (mr *multiReader[T]) Read(p []T) (n int, err error) {
-	for len(mr.readers) > 0 {
-		// Optimization to flatten nested multiReaders (Issue 13558).
-		if len(mr.readers) == 1 {
-			if r, ok := mr.readers[0].(*multiReader[T]); ok {
-				mr.readers = r.readers
-				continue
-			}
-		}
-		n, err = mr.readers[0].Read(p)
-		if err == EOF {
-			// Use eofReader instead of nil to avoid nil panic
-			// after performing flatten (Issue 18232).
-			mr.readers[0] = eofReader[T]{} // permit earlier GC
-			mr.readers = mr.readers[1:]
-		}
-		if n > 0 || err != EOF {
-			if err == EOF && len(mr.readers) > 0 {
-				// Don't return EOF yet. More readers remain.
-				err = nil
-			}
-			return
-		}
-	}
-	return 0, EOF
-}
-
-// MultiReader returns a Reader that's the logical concatenation of
-// the provided input readers. They're read sequentially. Once all
-// inputs have returned EOF, Read will return EOF.  If any of the readers
-// return a non-nil, non-EOF error, Read will return that error.
-func MultiReader[T any](readers ...Reader[T]) Reader[T] {
-	r := make([]Reader[T], len(readers))
-	copy(r, readers)
-	return &multiReader[T]{r}
-}
-
-type multiWriter[T any] struct {
-	writers []Writer[T]
-}
-
-func (t *multiWriter[T]) Write(p []T) (n int, err error) {
-	for _, w := range t.writers {
-		n, err = w.Write(p)
-		if err != nil {
-			return
-		}
-		if n != len(p) {
-			err = ErrShortWrite
-			return
-		}
-	}
-	return len(p), nil
-}
-
-// MultiWriter creates a writer that duplicates its writes to all the
-// provided writers, similar to the Unix tee(1) command.
-//
-// Each write is written to each listed writer, one at a time.
-// If a listed writer returns an error, that overall write operation
-// stops and returns the error; it does not continue down the list.
-func MultiWriter[T any](writers ...Writer[T]) Writer[T] {
-	allWriters := make([]Writer[T], 0, len(writers))
-	for _, w := range writers {
-		if mw, ok := w.(*multiWriter[T]); ok {
-			allWriters = append(allWriters, mw.writers...)
-		} else {
-			allWriters = append(allWriters, w)
-		}
-	}
-	return &multiWriter[T]{allWriters}
-}
