+++ _content/tour/generics/multi-type-params/example1.go
@@ -1,15 +1,21 @@
+//go:build OMIT
+// +build OMIT
+
+// All material is licensed under the Apache License Version 2.0, January 2004
+// http://www.apache.org/licenses/LICENSE-2.0
+
+// Sample program to show how to write a generic functions tha take multiple
+// generic parameters.
 package main
 
 import (
 	"fmt"
 )
 
-// =============================================================================
-
 // Print is a function that accepts a slice of some type L and a slice of some
-// type V (to be determined later). Each value of the labels slice will be joined
-// with the vals slice at the same index position. This code shows how the
-// generics type list can contain more than just one generic type and have
+// type V (to be determined later). Each value of the labels slice will be
+// joined with the vals slice at the same index position. This code shows how
+// the generics type list can contain more than just one generic type and have
 // different constraints for each.
 
 func Print[L any, V fmt.Stringer](labels []L, vals []V) {
@@ -17,8 +23,6 @@ func Print[L any, V fmt.Stringer](labels []L, vals []V) {
 		fmt.Println(labels[i], v.String())
 	}
 }
-
-// =============================================================================
 
 // This code defines a concrete type named user that implements the fmt.Stringer
 // interface. The String method just returns the name field from the user type.
@@ -35,6 +39,7 @@ func (u user) String() string {
 
 func main() {
 	labels := []int{1, 2, 3}
-	names := []user{{"bill"},{"jill"},{"joan"}}
+	names := []user{{"bill"}, {"jill"}, {"joan"}}
+
 	Print(labels, names)
 }
